requirements.txt
# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Azure SDK - CORRECTED VERSIONS
azure-identity==1.15.0
azure-mgmt-resource==23.1.0
azure-mgmt-storage==21.1.0
azure-mgmt-compute==30.3.0
azure-mgmt-network==25.1.0
azure-mgmt-monitor==6.0.0
azure-mgmt-sql==4.0.0b24      # Changed from 4.0.0 to 4.0.0b24
azure-mgmt-security==6.0.0
azure-mgmt-keyvault==10.3.0
azure-storage-blob==12.19.0

# API & Web
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
streamlit==1.28.1

# PDF Generation
reportlab==4.0.9

# Data Processing
pandas==2.1.3
numpy==1.24.3

# Visualization
plotly==5.18.0

# Utils
python-dotenv==1.0.0
requests==2.31.0
